requirements.txt
asgiref==3.7.2
Django==4.1.13
django-taggit==5.0.1
pillow==10.2.0
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2
